LayoutTests/imported/w3c/web-platform-tests/preload/single-download-preload-expected.txt

CONSOLE MESSAGE: <link rel=preload> must have a valid `as` value
CONSOLE MESSAGE: <link rel=preload> must have a valid `as` value
CONSOLE MESSAGE: The resource http://localhost:8800/preload/resources/square.png?background&single-download-preload was preloaded using link preload but not used within a few seconds from the window's load event. Please make sure it wasn't preloaded for nothing.
CONSOLE MESSAGE: The resource http://localhost:8800/preload/resources/white.mp4?single-download-preload was preloaded using link preload but not used within a few seconds from the window's load event. Please make sure it wasn't preloaded for nothing.
CONSOLE MESSAGE: The resource http://localhost:8800/preload/resources/sound_5.oga?single-download-preload was preloaded using link preload but not used within a few seconds from the window's load event. Please make sure it wasn't preloaded for nothing.
CONSOLE MESSAGE: The resource http://localhost:8800/preload/resources/foo.vtt?single-download-preload was preloaded using link preload but not used within a few seconds from the window's load event. Please make sure it wasn't preloaded for nothing.
  PASS - this text is here just so that the browser will download the font.

PASS Makes sure that preloaded resources are not downloaded again when used
